move decodeMoves and move2gtpvertex shared functions to /utils
