+++ game.js
@@ -1,10 +1,11 @@
 // vim: tw=120 softtabstop=4 shiftwidth=4
 
-const Bot = require('./bot').Bot;
-const decodeMoves = require('./bot').decodeMoves;
-const move2gtpvertex = require('./bot').move2gtpvertex;
+const { decodeMoves }    = require("./utils/decodeMoves");
+const { move2gtpvertex } = require("./utils/move2gtpvertex");
+
+const Bot     = require('./bot').Bot;
 const console = require('./console').console;
-const config = require('./config');
+const config  = require('./config');
 
 /**********/
 /** Game **/
